fix logging regression
