+++ src/py42/__version__.py
@@ -1,3 +1,3 @@
 # py42
 
-__version__ = "1.5.0"
+__version__ = "1.5.1"
