+++ tests/sdk/settings/test_debug.py
@@ -2,7 +2,7 @@ import pytest
 import logging
 
 import py42
-from py42.settings import debug, DebugSettings
+from py42.settings import debug
 
 
 @pytest.fixture
@@ -40,6 +40,7 @@ def trace_enabled():
 
 test_logger_name = "test"
 
+
 @pytest.fixture
 def custom_logger():
     default_logger = py42.settings.debug.logger
@@ -51,32 +52,41 @@ def custom_logger():
 def test_setting_debug_level_to_warning_sets_default_logger_to_warning(warning_enabled):
     assert debug.logger.level == logging.WARNING
 
+
 def test_setting_debug_level_to_info_sets_default_logger_to_info(info_enabled):
     assert debug.logger.level == logging.INFO
+
 
 def test_setting_debug_level_to_debug_sets_default_logger_to_debug(debug_enabled):
     assert debug.logger.level == logging.DEBUG
 
+
 def test_setting_debug_level_to_trace_sets_default_logger_to_debug(debug_enabled):
     assert debug.logger.level == logging.DEBUG
 
-def test_setting_debug_level_to_warning_sets_custom_logger_to_warning(custom_logger, warning_enabled):
+
+def test_setting_debug_level_to_warning_sets_custom_logger_to_warning(
+    custom_logger, warning_enabled
+):
     assert debug.logger.name == test_logger_name
     assert debug.logger.level == logging.WARNING
+
 
 def test_setting_debug_level_to_info_sets_custom_logger_to_info(custom_logger, info_enabled):
     assert debug.logger.name == test_logger_name
     assert debug.logger.level == logging.INFO
 
+
 def test_setting_debug_level_to_debug_sets_custom_logger_to_debug(custom_logger, debug_enabled):
     assert debug.logger.name == test_logger_name
     assert debug.logger.level == logging.DEBUG
+
 
 def test_setting_debug_level_to_trace_sets_custom_logger_to_debug(custom_logger, debug_enabled):
     assert debug.logger.name == test_logger_name
     assert debug.logger.level == logging.DEBUG
 
+
 def test_debug_settings_class_creates_default_logger():
-    test_debug = DebugSettings()
-    assert test_debug.logger.name == "py42"
-    assert test_debug.level == logging.NOTSET
+    assert debug.logger.name == "py42"
+    assert debug.level == logging.NOTSET
